# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental API",
        "description": "Dokumentacja systemu wypożyczalni",
        "version": "1.0.0"
    },
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Pobierz dane profilu",
                "operationId": "getUserProfile",
                "responses": {
                    "200": {
                        "description": "Zwraca dane zalogowanego użytkownika"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Zaktualizuj dane profilu",
                "operationId": "updateUserProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jan Kowalski"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dane zaktualizowane"
                    },
                    "422": {
                        "description": "Błąd walidacji"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/redeem-code": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Zrealizuj kod promocyjny (wpisywany ręcznie)",
                "operationId": "redeemPromoCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "BONUS2024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Konto doładowane"
                    },
                    "404": {
                        "description": "Nieprawidłowy kod"
                    },
                    "422": {
                        "description": "Kod wygasł lub użyty"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/codes": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Lista kodów przypisanych do użytkownika",
                "operationId": "getUserPromoCodes",
                "responses": {
                    "200": {
                        "description": "Lista kodów"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/codes/{id}/use": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Zrealizuj kod przypisany do konta (przez ID)",
                "operationId": "useAssignedPromoCode",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Konto doładowane"
                    },
                    "403": {
                        "description": "Kod nie należy do użytkownika"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}